Update 00_autobus.pnml
There. This is not going to make the busses show up on it's own but it WILL make them show up if someone makes the OAbG tramtype (optically-guided tram) and OAbS roadtype (municipal smartgrid for self-driving cars) according to the Standardized Roadtype Scheme as shown on October 24th, 2024 at 0148 UTC.

Read the comments I'll add to the roadtypetable for more information.